Move SortBy Test to a file
